Add a method add_nas_share_to_host to the Physical class
The add_nas_share_to_host method, given hostname, share_type (NFS or
SMB) and an export point POSTs to /host/share to add the new share to
Rubrik. This is done idempotently - if the export point already exists
as a share, the function does not POST and returns a string.

Also fix a small typo in the docstring of the Physical class.